# openapi.yaml
openapi: 3.0.0
paths:
...
  /user/comment:
    post:
      operationId: CommentController_createComment
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentDto'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/CommentResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Create new comment
      tags: &ref_3
        - Comment
  /user/comment/{commentId}:
    get:
      operationId: CommentController_getComment
      parameters:
        - name: commentId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/CommentResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Get comment by ID
      tags: *ref_3
    patch:
      operationId: CommentController_updateComment
      parameters:
        - name: commentId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentDto'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/CommentResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Update comment
      tags: *ref_3
    delete:
      operationId: CommentController_deleteComment
      parameters:
        - name: commentId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: object
                    nullable: true
                    example: null
                    description: No data returned
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Delete comment
      tags: *ref_3
  /user/comment/post/{postId}:
    get:
      operationId: CommentController_getCommentsByPost
      parameters:
        - name: postId
          required: true
          in: path
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            example: 20
            type: number
      responses:
        '200':
          description: Successful paginated response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CommentResponseDto'
                    description: Array of data items
                  paging:
                    type: object
                    description: Pagination metadata
                    properties:
                      total:
                        type: number
                        example: 100
                        description: Total number of items
                      page:
                        type: number
                        example: 1
                        description: Current page number
                      limit:
                        type: number
                        example: 10
                        description: Items per page
                      totalPages:
                        type: number
                        example: 10
                        description: Total number of pages
                    required:
                      - total
                      - page
                      - limit
                      - totalPages
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
                  - data
                  - paging
      summary: Get comments by post
      tags: *ref_3
  /user/comment/{commentId}/replies:
    get:
      operationId: CommentController_getReplies
      parameters:
        - name: commentId
          required: true
          in: path
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            example: 20
            type: number
      responses:
        '200':
          description: Successful paginated response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CommentResponseDto'
                    description: Array of data items
                  paging:
                    type: object
                    description: Pagination metadata
                    properties:
                      total:
                        type: number
                        example: 100
                        description: Total number of items
                      page:
                        type: number
                        example: 1
                        description: Current page number
                      limit:
                        type: number
                        example: 10
                        description: Items per page
                      totalPages:
                        type: number
                        example: 10
                        description: Total number of pages
                    required:
                      - total
                      - page
                      - limit
                      - totalPages
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
                  - data
                  - paging
      summary: Get replies to a comment
      tags: *ref_3
  /user/comment/author/{authorId}:
    get:
      operationId: CommentController_getCommentsByAuthor
      parameters:
        - name: authorId
          required: true
          in: path
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            example: 20
            type: number
      responses:
        '200':
          description: Successful paginated response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CommentResponseDto'
                    description: Array of data items
                  paging:
                    type: object
                    description: Pagination metadata
                    properties:
                      total:
                        type: number
                        example: 100
                        description: Total number of items
                      page:
                        type: number
                        example: 1
                        description: Current page number
                      limit:
                        type: number
                        example: 10
                        description: Items per page
                      totalPages:
                        type: number
                        example: 10
                        description: Total number of pages
                    required:
                      - total
                      - page
                      - limit
                      - totalPages
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
                  - data
                  - paging
      summary: Get comments by author
      tags: *ref_3
...
info:
  title: API Docs
  description: Custom Swagger UI in NestJS
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CommentResponseDto:
      type: object
      properties:
        id:
          type: string
          example: comment_abc123
        postId:
          type: string
          example: post_xyz789
        authorId:
          type: string
          example: user_def456
        parentCommentId:
          type: object
          example: comment_parent123
          nullable: true
        content:
          type: string
          example: This is a great post!
        replyCount:
          type: number
          example: 5
        isDeleted:
          type: boolean
          example: false
        createdAt:
          format: date-time
          type: string
          example: '2024-01-15T08:30:00Z'
        updatedAt:
          format: date-time
          type: string
          example: '2024-01-20T14:45:00Z'
      required:
        - id
        - postId
        - authorId
        - parentCommentId
        - content
        - replyCount
        - isDeleted
        - createdAt
        - updatedAt
    CreateCommentDto:
      type: object
      properties: {}
    UpdateCommentDto:
      type: object
      properties: {}
...